+++ table/plain_table_builder.cc
@@ -47,9 +47,10 @@ Status WriteBlock(
 }  // namespace
 
 // kPlainTableMagicNumber was picked by running
-//    echo rocksdb.plain.table | sha1sum
+//    echo rocksdb.table.plain | sha1sum
 // and taking the leading 64 bits.
-extern const uint64_t kPlainTableMagicNumber = 0x4f3418eb7a8f13b8ull;
+extern const uint64_t kPlainTableMagicNumber = 0x8242229663bf9564ull;
+extern const uint64_t kLegacyPlainTableMagicNumber = 0x4f3418eb7a8f13b8ull;
 
 PlainTableBuilder::PlainTableBuilder(const Options& options,
                                      WritableFile* file,
@@ -180,7 +181,8 @@ Status PlainTableBuilder::Finish() {
   }
 
   // Write Footer
-  Footer footer(kPlainTableMagicNumber);
+  // no need to write out new footer if we're using default checksum
+  Footer footer(kLegacyPlainTableMagicNumber);
   footer.set_metaindex_handle(metaindex_block_handle);
   footer.set_index_handle(BlockHandle::NullBlockHandle());
   std::string footer_encoding;